+++ src/Conversion.cpp
@@ -563,5 +563,99 @@ namespace ColorSpace {
 	  
 	  LuvConverter::ToColor(color, &luv);
 	}
+	
+	// Using values reported at https://bottosson.github.io/posts/oklab/#converting-from-linear-srgb-to-oklab
+	// instead of going through xyz. This ensures any whitepoint is ignored
+	void OkLabConverter::ToColorSpace(Rgb *color, OkLab *item) {
+	  if (!color->valid) {
+	    item->valid = false;
+	    return;
+	  }
+	  item->valid = true;
+	  double r = color->r / 255.0;
+	  double g = color->g / 255.0;
+	  double b = color->b / 255.0;
+	  
+	  r = ((r > 0.04045) ? std::pow((r + 0.055) / 1.055, 2.4) : (r / 12.92));
+	  g = ((g > 0.04045) ? std::pow((g + 0.055) / 1.055, 2.4) : (g / 12.92));
+	  b = ((b > 0.04045) ? std::pow((b + 0.055) / 1.055, 2.4) : (b / 12.92));
+	  
+	  double l = 0.4121656120 * r + 0.5362752080 * g + 0.0514575653 * b;
+	  double m = 0.2118591070 * r + 0.6807189584 * g + 0.1074065790 * b;
+	  double s = 0.0883097947 * r + 0.2818474174 * g + 0.6302613616 * b;
+	  
+	  
+	  l = std::cbrt(l);
+	  m = std::cbrt(m);
+	  s = std::cbrt(s);
+	  
+	  item->l = 0.2104542553 * l + 0.7936177850 * m - 0.0040720468 * s;
+	  item->a = 1.9779984951 * l - 2.4285922050 * m + 0.4505937099 * s;
+	  item->b = 0.0259040371 * l + 0.7827717662 * m - 0.8086757660 * s;
+	}
+	void OkLabConverter::ToColor(Rgb *color, OkLab *item) {
+	  if (!item->valid) {
+	    color->valid = false;
+	    return;
+	  }
+	  color->valid = true;
+	  double l = item->l + 0.3963377774 * item->a + 0.2158037573 * item->b;
+	  double m = item->l - 0.1055613458 * item->a - 0.0638541728 * item->b;
+	  double s = item->l - 0.0894841775 * item->a - 1.2914855480 * item->b;
+	  
+	  l = l * l * l;
+	  m = m * m * m;
+	  s = s * s * s;
+	  
+	  double r =  4.0767245293 * l - 3.3072168827 * m + 0.2307590544 * s;
+	  double g = -1.2681437731 * l + 2.6093323231 * m - 0.3411344290 * s;
+	  double b = -0.0041119885 * l - 0.7034763098 * m + 1.7068625689 * s;
+	  
+	  color->r = ((r > 0.0031308) ? (1.055*std::pow(r, 1 / 2.4) - 0.055) : (12.92*r)) * 255.0;
+	  color->g = ((g > 0.0031308) ? (1.055*std::pow(g, 1 / 2.4) - 0.055) : (12.92*g)) * 255.0;
+	  color->b = ((b > 0.0031308) ? (1.055*std::pow(b, 1 / 2.4) - 0.055) : (12.92*b)) * 255.0;
+	}
+	
+	void OkLchConverter::ToColorSpace(Rgb *color, OkLch *item) {
+	  if (!color->valid) {
+	    item->valid = false;
+	    return;
+	  }
+	  item->valid = true;
+	  OkLab lab;
+	  
+	  OkLabConverter::ToColorSpace(color, &lab);
+	  double l = lab.l;
+	  double c = std::sqrt(lab.a*lab.a + lab.b*lab.b);
+	  double h = std::atan2(lab.b, lab.a);
+	  
+	  h = h / M_PI * 180;
+	  if (h < 0) {
+	    h += 360;
+	  }
+	  else if (h >= 360) {
+	    h -= 360;
+	  }
+	  
+	  item->l = l;
+	  item->c = c;
+	  item->h = h;
+	}
+	void OkLchConverter::ToColor(Rgb *color, OkLch *item) {
+	  if (!item->valid) {
+	    color->valid = false;
+	    return;
+	  }
+	  color->valid = true;
+	  OkLab lab;
+	  
+	  item->h = item->h * M_PI / 180;
+	  
+	  lab.l = item->l;
+	  lab.a = std::cos(item->h)*item->c;
+	  lab.b = std::sin(item->h)*item->c;
+	  
+	  OkLabConverter::ToColor(color, &lab);
+	}
 }