+++ src/farver.cpp
@@ -87,6 +87,8 @@ SEXP convert_dispatch_to(SEXP colour, int to, SEXP white_from, SEXP white_to) {
     case XYZ: return convert_dispatch_impl<From, ColorSpace::Xyz>(colour, white_from, white_to);
     case YXY: return convert_dispatch_impl<From, ColorSpace::Yxy>(colour, white_from, white_to);
     case HCL: return convert_dispatch_impl<From, ColorSpace::Hcl>(colour, white_from, white_to);
+    case OKLAB: return convert_dispatch_impl<From, ColorSpace::OkLab>(colour, white_from, white_to);
+    case OKLCH: return convert_dispatch_impl<From, ColorSpace::OkLch>(colour, white_from, white_to);
   }
   
   // never happens
@@ -109,6 +111,8 @@ SEXP convert_dispatch_from(SEXP colour, int from, int to, SEXP white_from, SEXP 
     case XYZ: return convert_dispatch_to<ColorSpace::Xyz>(colour, to, white_from, white_to);
     case YXY: return convert_dispatch_to<ColorSpace::Yxy>(colour, to, white_from, white_to);
     case HCL: return convert_dispatch_to<ColorSpace::Hcl>(colour, to, white_from, white_to);
+    case OKLAB: return convert_dispatch_to<ColorSpace::OkLab>(colour, to, white_from, white_to);
+    case OKLCH: return convert_dispatch_to<ColorSpace::OkLch>(colour, to, white_from, white_to);
   }
   
   // never happens so we just return the input to quiet the compiler
@@ -229,6 +233,8 @@ SEXP compare_dispatch_to(SEXP from, SEXP to, int to_space, int dist, bool sym, S
     case XYZ: return compare_dispatch_impl<From, ColorSpace::Xyz>(from, to, dist, sym, white_from, white_to);
     case YXY: return compare_dispatch_impl<From, ColorSpace::Yxy>(from, to, dist, sym, white_from, white_to);
     case HCL: return compare_dispatch_impl<From, ColorSpace::Hcl>(from, to, dist, sym, white_from, white_to);
+    case OKLAB: return compare_dispatch_impl<From, ColorSpace::OkLab>(from, to, dist, sym, white_from, white_to);
+    case OKLCH: return compare_dispatch_impl<From, ColorSpace::OkLch>(from, to, dist, sym, white_from, white_to);
   }
   
   // never happens
@@ -249,6 +255,8 @@ SEXP compare_dispatch_from(SEXP from, SEXP to, int from_space, int to_space, int
     case RGB: return compare_dispatch_to<ColorSpace::Rgb>(from, to, to_space, dist, sym, white_from, white_to);
     case XYZ: return compare_dispatch_to<ColorSpace::Xyz>(from, to, to_space, dist, sym, white_from, white_to);
     case YXY: return compare_dispatch_to<ColorSpace::Hcl>(from, to, to_space, dist, sym, white_from, white_to);
+    case OKLAB: return compare_dispatch_to<ColorSpace::OkLab>(from, to, to_space, dist, sym, white_from, white_to);
+    case OKLCH: return compare_dispatch_to<ColorSpace::OkLch>(from, to, to_space, dist, sym, white_from, white_to);
   }
   
   // never happens so we just return the input to quiet the compiler